# Backend/Ticketing.Ticket/src/Ticketing.Ticket.Application/Dtos/Responses/TicketDetailResponse.cs
namespace Ticketing.Ticket.Application.Dtos.Responses;
public sealed record TicketDetailResponse
{
  public Guid Id { get; set; }
  public string Subject { get; set; } = string.Empty;
  public string Description { get; set; } = string.Empty;
  public string Status { get; set; } = string.Empty;
  public Guid UserId { get; set; }
  public List<TicketReplyResponse> Replies { get; set; } = new();
}